Added Wrapper for annotations to allow simpler reading of annotations.
This is quite useful if there is a class based attribute in the annotation, which is often the case.

Reading of attribute values will be similar to use the annotation type directly!

Class based attributes will then be accessible as TypeMirror, FQN String ot TypeMirrorWrapper.

Embedded annotations can be configured to be automatically wrapped as well if not explicitly configured added at the AnnotationWrapper annotation.