øøAdded enhanced support for repeatable annotation processing and a Compile Message Processor to generate compiler message classes
This includes:
- automatic support of repeatable wrapper annotations
- method to get all annotated elements (annotation + repeatable wrapper annotations)
- Special Case: Repeatable annotation wrapper - wrap returns List of wrappers

In general no explicit handling of the repeatable wrapper annotation is needed anymore :)

The Compile Message Processor allows generation of compiler message enum at compile time.
Compiler messages can now be declared via annotation on the type or methods where they are used.
This is kind of an experimental feature. It has to show how it feels in praxis.